Relocate experimental vector classes to a vector directory
Add an implementation of the functional vector envs
