+++ gymnasium/experimental/vector/__init__.py
@@ -1,23 +1,13 @@
-"""Root __init__ of the gym experimental wrappers."""
-
-
-from gymnasium.experimental import functional, wrappers
-from gymnasium.experimental.functional import FuncEnv
+"""Experimental vector env API."""
 from gymnasium.experimental.vector.async_vector_env import AsyncVectorEnv
 from gymnasium.experimental.vector.sync_vector_env import SyncVectorEnv
 from gymnasium.experimental.vector.vector_env import VectorEnv, VectorWrapper
 
 
 __all__ = [
-    # Functional
-    "FuncEnv",
-    "functional",
-    # Wrappers
-    "wrappers",
     # Vector
     "VectorEnv",
     "VectorWrapper",
     "SyncVectorEnv",
     "AsyncVectorEnv",
-    # "vector",
 ]
