+++ gymnasium/experimental/vector/async_vector_env.py
@@ -18,7 +18,7 @@ from gymnasium.error import (
     CustomSpaceError,
     NoAsyncCallError,
 )
-from gymnasium.experimental.vector_env import VectorEnv
+from gymnasium.experimental.vector.vector_env import VectorEnv
 from gymnasium.vector.utils import (
     CloudpickleWrapper,
     batch_space,
