+++ tests/vector/test_async_vector_env.py
@@ -9,7 +9,7 @@ from gymnasium.error import (
     ClosedEnvironmentError,
     NoAsyncCallError,
 )
-from gymnasium.experimental.async_vector_env import AsyncVectorEnv
+from gymnasium.experimental.vector import AsyncVectorEnv
 from gymnasium.spaces import Box, Discrete, MultiDiscrete, Tuple
 from tests.vector.utils import (
     CustomSpace,
