+++ tests/vector/test_sync_vector_env.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from gymnasium.envs.registration import EnvSpec
-from gymnasium.experimental.sync_vector_env import SyncVectorEnv
+from gymnasium.experimental.vector import SyncVectorEnv
 from gymnasium.spaces import Box, Discrete, MultiDiscrete, Tuple
 from tests.envs.utils import all_testing_env_specs
 from tests.vector.utils import (
